Fix tests, improve test coverage.
